Add a function that saves a cropped point cloud
In order to extract a portion of the point cloud available on the scene, savePointCloud function was added.
It saves the object under the name "cropped pointcloud" that can be extracted in many ways.
Our specific need is in obtaining the point cloud of an object that we do not dispose of and saving it for later use.

signed-by: mariem.khlifi@tum.de