Generate bounding boxes files for YOLO
Together with the poses files, we calculate the bounding boxes in a separate file following the covention mentioned in https://github.com/AlexeyAB/darknet#how-to-train-to-detect-your-custom-objects.
For every declared object and for every frame, we save the information about the bounding boxes.

This is necessary to fine-tune YOLO with our own dataset.

signed-by: mariem.khlifi@tum.de